refactors state related classes closes #1721
excludes index* from code coverage

removes changes to babelrc and plugin export-default-from
