+++ __tests__/src/actions/window.test.js
@@ -1,5 +1,5 @@
-import * as actions from '../../../src/actions/index';
-import ActionTypes from '../../../src/action-types';
+import * as actions from '../../../src/state/actions';
+import ActionTypes from '../../../src/state/actions/action-types';
 
 describe('window actions', () => {
   describe('addWindow', () => {
