+++ __tests__/src/components/ManifestMetadata.test.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { actions, store } from '../../../src/store';
+import createStore from '../../../src/state/createStore';
+import * as actions from '../../../src/state/actions';
 import ManifestMetadata from '../../../src/components/ManifestMetadata';
 import fixture from '../../fixtures/version-2/002.json';
 
 describe('ManifestMetadata', () => {
   let wrapper;
   let manifest;
+  const store = createStore();
   beforeEach(() => {
     store.dispatch(actions.receiveManifest('foo', fixture));
     manifest = store.getState().manifests.foo;
