+++ __tests__/src/components/WindowSideBar.test.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { store } from '../../../src/store';
+import createStore from '../../../src/state/createStore';
 import { WindowSideBar } from '../../../src/components/WindowSideBar';
 
 describe('WindowSideBar', () => {
   let wrapper;
+  const store = createStore();
   beforeEach(() => {
     wrapper = shallow(<WindowSideBar store={store} windowId="1" classes={{}} />);
   });
