+++ __tests__/src/components/WindowTopBarButtons.test.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { store } from '../../../src/store';
+import createStore from '../../../src/state/createStore';
 import { WindowTopBarButtons } from '../../../src/components/WindowTopBarButtons';
 
 describe('WindowTopBarButtons', () => {
   let wrapper;
+  const store = createStore();
   beforeEach(() => {
     wrapper = shallow(<WindowTopBarButtons store={store} />);
   });
