+++ __tests__/src/components/WorkspaceControlPanelButtons.test.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { store } from '../../../src/store';
+import createStore from '../../../src/state/createStore';
 import { WorkspaceControlPanelButtons } from '../../../src/components/WorkspaceControlPanelButtons';
 
 describe('WorkspaceControlPanelButtons', () => {
   let wrapper;
+  const store = createStore();
   beforeEach(() => {
     wrapper = shallow(<WorkspaceControlPanelButtons store={store} />);
   });
