+++ __tests__/src/reducers/manifests.test.js
@@ -1,5 +1,5 @@
-import reducer from '../../../src/reducers/manifests';
-import ActionTypes from '../../../src/action-types';
+import reducer from '../../../src/state/reducers/manifests';
+import ActionTypes from '../../../src/state/actions/action-types';
 
 describe('manifests reducer', () => {
   it('should handle REQUEST_MANIFEST', () => {
