+++ jest.json
@@ -1,7 +1,8 @@
 {
   "collectCoverage": true,
   "collectCoverageFrom": [
-    "src/**/*.{js,jsx}"
+    "src/**/*.{js,jsx}",
+    "!src/**/index*.js"
   ],
   "coverageDirectory": "<rootDir>/coverage",
   "coverageReporters": ["html", "lcov"],
