+++ src/components/ManifestListItem.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { actions } from '../store';
+import * as actions from '../state/actions';
 import ns from '../config/css-ns';
