+++ src/index-core.js
@@ -1,0 +1,5 @@
+import createStore from './state/createStore';
+import * as actions from './state/actions';
+
+const store = createStore();
+export { store, actions };
