+++ src/state/reducers/config.js
@@ -1,5 +1,5 @@
 import deepmerge from 'deepmerge';
-import ActionTypes from '../action-types';
+import ActionTypes from '../actions/action-types';
 
 /**
  * configReducer - does a deep merge of the config
