+++ src/state/reducers/manifests.js
@@ -1,5 +1,5 @@
 import manifesto from 'manifesto.js';
-import ActionTypes from '../action-types';
+import ActionTypes from '../actions/action-types';
 
 /**
  * manifestsReducer
