+++ src/state/reducers/windows.js
@@ -1,4 +1,4 @@
-import ActionTypes from '../action-types';
+import ActionTypes from '../actions/action-types';
 
 /**
  * windowsReducer
